+++ .github/workflows/php.yml
@@ -2,9 +2,9 @@ name: PHP Composer
 
 on:
   push:
-    branches: [main]
+    branches: [main,develop]
   pull_request:
-    branches: [main]
+    branches: [main,develop]
   workflow_dispatch: ~
 
 jobs:
