+++ src/.travis.yml
@@ -10,7 +10,7 @@ notifications:
 
 branches:
   only:
-    - master
+    - main
 
 cache:
   directories:
