+++ src/composer.json
@@ -38,11 +38,5 @@
             "composer/installers": true,
             "dealerdirect/phpcodesniffer-composer-installer": true
         }
-    },
-    "repositories": {
-        "alma/alma-php-client": {
-            "type": "path",
-            "url": "/app/alma/alma-php-client"
-        }
     }
 }
