+++ src/includes/exceptions/class-alma-api-create-payments-exception.php
@@ -2,7 +2,7 @@
 /**
  * Alma_Api_Create_Payments_Exception.
  *
- * @since 4.0.0
+ * @since 5.0.0
  *
  * @package Alma_Gateway_For_Woocommerce
  * @subpackage Alma_Gateway_For_Woocommerce/includes/exceptions
@@ -22,18 +22,19 @@ if ( ! defined( 'ABSPATH' ) ) {
  */
 class Alma_Api_Create_Payments_Exception extends Alma_Exception {
 
-
 	/**
 	 * Constructor.
 	 *
 	 * @param string  $order_id The order id.
 	 * @param FeePlan $fee_plan The fee plans.
+	 * @param array   $payload The payload.
 	 */
-	public function __construct( $order_id, $fee_plan ) {
+	public function __construct( $order_id, $fee_plan, $payload ) {
 		$message = sprintf(
-			'Error while creating payment. Order id : %s, Plan definition : %s',
+			'Error while creating payment. Order id : "%s", Plan definition : "%s", Payload "%s"',
 			$order_id,
-			wp_json_encode( $fee_plan )
+			wp_json_encode( $fee_plan ),
+			wp_json_encode( $payload )
 		);
 
 		parent::__construct( $message );
