+++ src/includes/helpers/class-alma-constants-helper.php
@@ -20,18 +20,20 @@ if ( ! defined( 'ABSPATH' ) ) {
  */
 class Alma_Constants_Helper {
 
+	const GATEWAY_ID                 = 'alma';
+	const GATEWAY_ID_PAY_NOW         = 'alma_pay_now';
+	const GATEWAY_ID_PAY_LATER       = 'alma_pay_later';
+	const GATEWAY_ID_MORE_THAN_FOUR  = 'alma_pnx_plus_4';
+	const GATEWAY_ID_IN_PAGE         = 'alma_in_page';
+	const GATEWAY_ID_IN_PAGE_PAY_NOW = 'alma_in_page_pay_now';
 
-	const GATEWAY_ID         = 'alma';
-	const GATEWAY_ID_IN_PAGE = 'alma_in_page';
+	const JQUERY_CART_UPDATE_EVENT = 'updated_cart_totals';
 
 	const ALMA_PAYMENT_PLAN_TABLE_ID_TEMPLATE = 'alma-payment-plan-table-%s-installments';
 	const ALMA_PAYMENT_PLAN_TABLE_CSS_CLASS   = 'js-alma-payment-plan-table';
-	const ALMA_GATEWAY_PAY_LATER              = 'alma_pay_later';
-	const ALMA_GATEWAY_PAY_NOW                = 'alma_pay_now';
-	const ALMA_GATEWAY_PAY_MORE_THAN_FOUR     = 'alma_pnx_plus_4';
-	const JQUERY_CART_UPDATE_EVENT            = 'updated_cart_totals';
-	const PREFIX_REFUND_COMMENT               = 'Refund made via WooCommerce back-office - ';
-	const FLAG_ORDER_FULLY_REFUNDED           = 'alma_order_fully_refunded';
+
+	const PREFIX_REFUND_COMMENT     = 'Refund made via WooCommerce back-office - ';
+	const FLAG_ORDER_FULLY_REFUNDED = 'alma_order_fully_refunded';
 
 	const SUCCESS                        = 'success';
 	const REFUND_NOTICE_META_KEY         = 'alma_refund_notices';
@@ -43,24 +45,22 @@ class Alma_Constants_Helper {
 	const ALMA_FEE_PLAN                  = 'alma_fee_plan';
 	const ALMA_FEE_PLAN_IN_PAGE          = 'alma_fee_plan_in_page';
 	const DEFAULT_FEE_PLAN               = 'general_3_0_0';
+	const PAY_NOW_FEE_PLAN               = 'general_1_0_0';
 	const PAYMENT_METHOD                 = 'payment_method';
 	const DEFAULT_CHECK_VARIATIONS_EVENT = 'check_variations';
 	const AMOUNT_PLAN_KEY_REGEX          = '#^(min|max)_amount_general_[0-9]+_[0-9]+_[0-9]+$#';
 	const SORT_PLAN_KEY_REGEX            = '/^(general|pos)_([0-9]+)_([0-9]+)_([0-9]+)$/';
-	const GATEWAY_DESCRIPTION_IN_PAGE    = 'gateway_description_in_page';
-	const GATEWAY_TITLE                  = 'pay_with_alma';
-	const GATEWAY_TITLE_IN_PAGE          = 'pay_with_alma_in_page';
-	const PAYMENT_METHOD_PNX             = 'payment_method_pnx';
-	const PAYMENT_METHOD_PAY_LATER       = 'payment_method_pay_later';
-	const PAYMENT_METHOD_PAY_NOW         = 'payment_method_pay_now';
-	const PAYMENT_METHOD_PNX_PLUS_4      = 'payment_method_pnx_plus_4';
-	const NOTICE_NONCE_NAME              = 'wc_alma_notice_nonce';
-	const ALMA_LOGO_PATH                 = 'images/alma_logo.svg';
-	const ALMA_SHORT_LOGO_PATH           = 'images/alma_short_logo.svg';
-	const ALMA_PATH_CHECKOUT_JS          = 'js/alma-checkout.js';
-	const ALMA_PATH_CHECKOUT_IN_PAGE_JS          = 'js/alma-checkout-in-page.js';
-	const ALMA_PATH_CHECKOUT_CDN_IN_PAGE_JS  = 'https://cdn.jsdelivr.net/npm/@alma/in-page@2.x/dist/index.umd.js';
-	const ALMA_PATH_CHECKOUT_CSS         = 'css/alma-checkout.css';
+
+	const NOTICE_NONCE_NAME                 = 'wc_alma_notice_nonce';
+	const ALMA_LOGO_PATH                    = 'images/alma_logo.svg';
+	const ALMA_SHORT_LOGO_PATH              = 'images/alma_short_logo.svg';
+	const ALMA_PATH_CHECKOUT_JS             = 'js/alma-checkout.js';
+	const ALMA_PATH_CHECKOUT_IN_PAGE_JS     = 'js/alma-checkout-in-page.js';
+	const ALMA_PATH_CHECKOUT_CDN_IN_PAGE_JS = 'https://cdn.jsdelivr.net/npm/@alma/in-page@2.x/dist/index.umd.js';
+	const ALMA_PATH_CHECKOUT_CSS            = 'css/alma-checkout.css';
+	const PAY_IN_INSTALLMENTS               = 'Pay in installments';
+
+	const PAY_NOW = 'Pay now by credit card';
 
 	const ERROR = 'error';
 
@@ -74,5 +74,19 @@ class Alma_Constants_Helper {
 		'pending',
 		'failed',
 		'cancelled',
+	);
+
+	/**
+	 * The gateways.
+	 *
+	 * @var string[]
+	 */
+	public static $gateways_ids = array(
+		self::GATEWAY_ID,
+		self::GATEWAY_ID_IN_PAGE,
+		self::GATEWAY_ID_PAY_NOW,
+		self::GATEWAY_ID_IN_PAGE_PAY_NOW,
+		self::GATEWAY_ID_PAY_LATER,
+		self::GATEWAY_ID_MORE_THAN_FOUR,
 	);
 }
