+++ src/includes/helpers/class-alma-settings-helper.php
@@ -11,6 +11,9 @@
 
 namespace Alma\Woocommerce\Helpers;
 
+use Alma\Woocommerce\Alma_Settings;
+use Alma\Woocommerce\Exceptions\Alma_No_Credentials_Exception;
+
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
@@ -21,8 +24,6 @@ if ( ! defined( 'ABSPATH' ) ) {
  * Helps Settings to define defaults values (including in i18n context)
  */
 class Alma_Settings_Helper {
-
-
 
 	/**
 	 * Get default settings.
@@ -37,17 +38,18 @@ class Alma_Settings_Helper {
 			'payment_upon_trigger_display_text'          => 'at_shipping',
 			'selected_fee_plan'                          => Alma_Constants_Helper::DEFAULT_FEE_PLAN,
 			'enabled_general_3_0_0'                      => 'yes',
-			'title_pay_with_alma'                        => self::default_pay_with_alma(),
-			'title_pay_with_alma_in_page'                => self::default_pay_with_alma_in_page(),
-			'description_gateway_description_in_page'    => self::default_pay_with_alma_in_page_description(),
-			'title_payment_method_pnx'                   => self::default_pnx_title(),
-			'description_payment_method_pnx'             => self::default_payment_description(),
-			'title_payment_method_pay_now'               => self::default_pay_now_title(),
-			'description_payment_method_pay_now'         => self::default_pay_now_description(),
-			'title_payment_method_pay_later'             => self::default_pay_later_title(),
-			'description_payment_method_pay_later'       => self::default_payment_description(),
-			'title_payment_method_pnx_plus_4'            => self::default_pnx_plus_4_title(),
-			'description_payment_method_pnx_plus_4'      => self::default_payment_description(),
+			'title_alma_in_page'                         => self::default_pay_with_alma_in_page(),
+			'description_alma_in_page'                   => self::default_pay_with_alma_in_page_description(),
+			'title_alma_in_page_pay_now'                 => self::default_pay_with_alma_in_page_pay_now(),
+			'description_alma_in_page_pay_now'           => self::default_pay_with_alma_in_page_pay_now_description(),
+			'title_alma'                                 => self::default_pnx_title(),
+			'description_alma'                           => self::default_payment_description(),
+			'title_alma_pay_now'                         => self::default_pay_now_title(),
+			'description_alma_pay_now'                   => self::default_pay_now_description(),
+			'title_alma_pay_later'                       => self::default_pay_later_title(),
+			'description_alma_pay_later'                 => self::default_payment_description(),
+			'title_alma_pnx_plus_4'                      => self::default_pnx_plus_4_title(),
+			'description_alma_pnx_plus_4'                => self::default_payment_description(),
 			'display_cart_eligibility'                   => 'yes',
 			'display_product_eligibility'                => 'yes',
 			'variable_product_price_query_selector'      => self::default_variable_price_selector(),
@@ -70,11 +72,11 @@ class Alma_Settings_Helper {
 	 *
 	 * @return string
 	 */
-	public static function default_pay_with_alma() {
+	public static function default_pay_with_alma_in_page() {
 		if ( Alma_Internationalization_Helper::is_site_multilingual() ) {
-			return 'Pay with Alma';
+			return Alma_Constants_Helper::PAY_IN_INSTALLMENTS;
 		}
-		return __( 'Pay with Alma', 'alma-gateway-for-woocommerce' );
+		return __( 'Pay in installments', 'alma-gateway-for-woocommerce' );
 	}
 
 	/**
@@ -82,11 +84,11 @@ class Alma_Settings_Helper {
 	 *
 	 * @return string
 	 */
-	public static function default_pay_with_alma_in_page() {
+	public static function default_pay_with_alma_in_page_pay_now() {
 		if ( Alma_Internationalization_Helper::is_site_multilingual() ) {
-			return 'Pay with Alma In Page';
+			return Alma_Constants_Helper::PAY_NOW;
 		}
-		return __( 'Pay with Alma In Page', 'alma-gateway-for-woocommerce' );
+		return __( 'Pay now by credit card', 'alma-gateway-for-woocommerce' );
 	}
 
 	/**
@@ -96,7 +98,7 @@ class Alma_Settings_Helper {
 	 */
 	public static function default_pnx_title() {
 		if ( Alma_Internationalization_Helper::is_site_multilingual() ) {
-			return 'Pay in installments';
+			return Alma_Constants_Helper::PAY_IN_INSTALLMENTS;
 		}
 		return __( 'Pay in installments', 'alma-gateway-for-woocommerce' );
 	}
@@ -108,7 +110,7 @@ class Alma_Settings_Helper {
 	 */
 	public static function default_pay_now_title() {
 		if ( Alma_Internationalization_Helper::is_site_multilingual() ) {
-			return 'Pay now by credit card';
+			return Alma_Constants_Helper::PAY_NOW;
 		}
 		return __( 'Pay now by credit card', 'alma-gateway-for-woocommerce' );
 	}
@@ -131,6 +133,19 @@ class Alma_Settings_Helper {
 	 *
 	 * @return string
 	 */
+	public static function default_pay_with_alma_in_page_pay_now_description() {
+		if ( Alma_Internationalization_Helper::is_site_multilingual() ) {
+			return 'Fast and secure payment by credit card.';
+		}
+
+		return __( 'Fast and secure payment by credit card.', 'alma-gateway-for-woocommerce' );
+	}
+
+	/**
+	 * Gets the default description for all payment methods (pnx, pnx+4, pay later).
+	 *
+	 * @return string
+	 */
 	public static function default_payment_description() {
 		if ( Alma_Internationalization_Helper::is_site_multilingual() ) {
 			return 'Fast and secure payment by credit card';
@@ -145,9 +160,9 @@ class Alma_Settings_Helper {
 	 */
 	public static function default_pay_with_alma_in_page_description() {
 		if ( Alma_Internationalization_Helper::is_site_multilingual() ) {
-			return 'Fast and secure payment by credit card *2 *3 *4';
+			return 'Fast and secure payment by credit card.';
 		}
-		return __( 'Fast and secure payment by credit card  *2 *3 *4', 'alma-gateway-for-woocommerce' );
+		return __( 'Fast and secure payment by credit card.', 'alma-gateway-for-woocommerce' );
 	}
 
 
@@ -216,6 +231,53 @@ class Alma_Settings_Helper {
 			return 'Some products cannot be paid with monthly or deferred installments';
 		}
 		return __( 'Some products cannot be paid with monthly or deferred installments', 'alma-gateway-for-woocommerce' );
+	}
+
+	/**
+	 *  Reset min and max amount for all plans.
+	 *
+	 * @param array $alma_settings The settings.
+	 *
+	 * @return mixed
+	 */
+	public static function reset_plans( $alma_settings ) {
+		foreach ( array_keys( $alma_settings ) as $key ) {
+			if ( Alma_Tools_Helper::is_amount_plan_key( $key ) ) {
+				$alma_settings[ $key ] = null;
+			}
+		}
+
+		$alma_settings['allowed_fee_plans'] = null;
+		$alma_settings['live_merchant_id']  = null;
+		$alma_settings['test_merchant_id']  = null;
+
+		return $alma_settings;
+	}
+
+	/**
+	 * Checks the api keys
+	 *
+	 * @param boolean $has_keys Keys.
+	 * @param bool    $throw_exception Do we want to throw the exception.
+	 *
+	 * @return void
+	 * @throws Alma_No_Credentials_Exception The exception.
+	 */
+	public static function check_alma_keys( $has_keys, $throw_exception = true ) {
+		// Do we have keys for the environment?
+		if ( ! $has_keys ) { // nope.
+			$message = sprintf(
+			// translators: %s: Admin settings url.
+				__( 'Alma is almost ready. To get started, <a href="%s">fill in your API keys</a>.', 'alma-gateway-for-woocommerce' ),
+				esc_url( Alma_Assets_Helper::get_admin_setting_url() )
+			);
+
+			alma_plugin()->admin_notices->add_admin_notice( 'no_alma_keys', 'notice notice-warning', $message );
+
+			if ( $throw_exception ) {
+				throw new Alma_No_Credentials_Exception( $message );
+			}
+		}
 	}
 
 }
