+++ src/public/templates/alma-checkout-plans-classic.php
@@ -9,10 +9,6 @@
  */
 
 ?>
-<h5 id="<?php echo esc_html( $id ); ?>_accordion" style="font-weight: 400; font-size: 1.1em;">
-	<?php echo esc_html( $title ); ?>
-</h5>
-
 <div>
 <?php echo $description; // phpcs:ignore WordPress.Security.EscapeOutput ?>
 <br>
