+++ src/public/templates/alma-checkout-plans-in-page.php
@@ -2,7 +2,7 @@
 /**
  * Template.
  *
- * @since 4.2.0
+ * @since 5.0.0
  *
  * @package Alma_Gateway_For_Woocommerce
  * @subpackage Alma_Gateway_For_Woocommerce/public/templates
