+++ src/public/templates/partials/alma-checkout-plan.php
@@ -16,7 +16,7 @@
 	id="<?php echo esc_attr( \Alma\Woocommerce\Helpers\Alma_Constants_Helper::GATEWAY_ID ); ?>_alma_fee_plan_<?php echo esc_attr( $plan_key ); ?>"
 	name="<?php echo esc_attr( \Alma\Woocommerce\Helpers\Alma_Constants_Helper::ALMA_FEE_PLAN ); ?>"
 	data-default="<?php echo $is_checked ? '1' : '0'; ?>"
-	style="<?php echo ( \Alma\Woocommerce\Helpers\Alma_Constants_Helper::ALMA_GATEWAY_PAY_NOW == $id ) ? 'margin: 18px -1px;' : 'margin-right: -1px;'; ?>; vertical-align: middle;"
+	style="<?php echo ( \Alma\Woocommerce\Helpers\Alma_Constants_Helper::GATEWAY_ID_PAY_NOW == $id ) ? 'margin: 18px -1px;' : 'margin-right: -1px;'; ?>; vertical-align: middle;"
 	<?php echo $is_checked ? 'checked' : ''; ?>
 	onchange="if (this.checked) { jQuery( '<?php echo esc_js( $plan_class ); ?>' ).hide(); jQuery(this).closest('li.wc_payment_method').find( '<?php echo esc_js( $plan_id ); ?>' ).show() }"
 >
@@ -27,7 +27,7 @@
 >
 
 	<?php
-	if ( \Alma\Woocommerce\Helpers\Alma_Constants_Helper::ALMA_GATEWAY_PAY_NOW == $id ) {
+	if ( \Alma\Woocommerce\Helpers\Alma_Constants_Helper::GATEWAY_ID_PAY_NOW == $id ) {
 		?>
 		<span class="logoContainer" style="margin-top: 10px !important;">
 		<img src="<?php echo esc_attr( $logo_url ); ?>"
